Fixed the problem when there were duplicated mappingFragments. Don't know why that happened.
Fixed a problem with duplicated query params. When I made this query:

foreach (Provincia provincia in provincias)
{
                EFBatchOperation.For(context, context.Set<Poblacion>()).Where(x => x.Codigo.StartsWith(provincia.Codigo)).Update(pob => pob.Provincia.Id, poblacion => provincia.Id);
}

That launches an exception with the message "The variable name '@p__linq__0' has already been declared. Variable names must be unique within a query batch or stored procedure." Debugging, I found that at the concatenation of parameters of the method Update there were two params called "p__linq__0" because at the 'queryInformation' var there was one parameter for the 'where' clause of the query and at 'mqueryInfo' var there was another for assigning the value. That caused the error. For fixing it, I renamed the duplicated param at the update clause and that 'fixes' the issue. I'm sure that it's not the right way to fix it, but I couldn't find another way.